Add Custom error page and Site map
Added 404.md to generate /404.html - Custom Page not found! error page
Modified _includes/header.html to use only the top_nav_name if it exists
 to determine if the page belongs on top navigation bar.
Modified about.md to use the top_nav_name to include it in the navigation bar.
Added robots.txt and sitemap.md for generated sitemap.xml file.
Added sitemap.html to generate a sitemap in html form.
Added a page title to the main page to be used on the new sitemap page.

I wanted the Custom 404 Error page generated with the default site template.
I wanted a title other than the site title but did not want the page.
I decided top_nav_name is a better variable name than short_title.
I included a top_nav_name in the sitemap.html so that it is included
 in the top navigation.