Added a man collection for the latest post
I was able to install hplip 3.20.6 from source on Linux Mint 20
The above post referenced man pages so I created a collection for
man pages.
I am having issues with the style for man pages but have not dealt
with it yet.  There are issues with jekyll style and the DT used
in man pages.

Included pan pages in the Pages dropdown navigation.